+++ src/website/app/demos/Card/examples/index.js
@@ -16,14 +16,20 @@
 
 /* @flow */
 import cardExamples from './card';
+import cardActionsExamples from './card-actions';
 import cardBlockExamples from './card-block';
+import cardDividerExamples from './card-divider';
 import cardImageExamples from './card-image';
+import cardStatusExamples from './card-status';
 import cardTitleExamples from './card-title';
 
 const examples = [].concat(
   cardExamples,
+  cardActionsExamples,
   cardBlockExamples,
+  cardDividerExamples,
   cardImageExamples,
+  cardStatusExamples,
   cardTitleExamples
 );