+++ src/website/app/demos/index.js
@@ -17,6 +17,7 @@
 /* @flow */
 import flatten from 'lodash/flatten';
 import createKeyMap from '../utils/createKeyMap';
+import avatar from './Avatar';
 import button from './Button';
 import card from './Card';
 import dropdown from './Dropdown';
@@ -25,10 +26,13 @@ import icon from './Icon';
 import link from './Link';
 import menu from './Menu';
 import popover from './Popover';
+import radio from './Radio';
+import textArea from './TextArea';
 import textInput from './TextInput';
 import themeProvider from './ThemeProvider';
 
 const demos = flatten([
+  avatar,
   button,
   card,
   dropdown,
@@ -37,6 +41,8 @@ const demos = flatten([
   link,
   menu,
   popover,
+  radio,
+  textArea,
   textInput,
   themeProvider
 ]);